ref(errors): Move event processor code to processing
